myproject/requirements.txt
Django==5.1.2
python-dotenv==1.0.1
jazzmin==2.5.1
mysqlclient==2.2.4
Pillow==10.4.0
